User.Organization()
Given a user with a valid OrgID, calling user.Organization() will return the organization that corresponds to that ID in the database. Works only by re-calling db.Init() which returns a storer object that we run this against, and I’d like to find a better/cleaner way to do that but haven’t so far (suggestions welcome).

+ Changes multiple int64s to ints (int64 was a pain in the ass to work with and unnecessary)
+ Formatting fixes
+ CreateNewUser now specifies the default value (NOW() in UTC) for created_at. It also doesn’t halt if the error in its first pass at finding an existing user is sql.ErrNoRows.
+ Multiple db.Select statements changed to db.Get. Select wants to populate a slice/array, not just one variable. Get is what you need for a single variable.